All tests passed on commit 32d9850, posting as v2.0

+++ test_scripts/_run_tests.bash
@@ -4,10 +4,10 @@ echo
 echo _run_tests.bash
 echo
 
-cd ../src
-rm -rf o/ ../bin/reseek*
-./build_linux_x86.bash
-cd ../test_scripts
+if [ x$reseek == x ] ; then
+	echo "reseek not set"
+	exit 1
+fi
 
 rm -rf ../test_output
 mkdir ../test_output
@@ -16,7 +16,7 @@ mkdir -p ../test_results
 log=../test_output/TEST_LOG.txt
 
 date=`date "+%Y-%m-%d/%H:%M:%S"`
-ver=`../bin/reseek --version | tr -d ' \n\r'`
+ver=`$reseek --version | tr -d ' \n\r'`
 echo $date $ver STARTED >> $log
 git status >> $log
